Create base rules definition class and derivatives for C# and VB.NET
Create a new abstract base class to provide SonarQube rule repositories and definitions based on issues that can be detected by the InspectCode command line tool. This base class is then used by derivatives for the languages C# and VisualBasic.NET in order to create specific rule repositories. Most of the base implementation has been extracted from the previous implementation of the CSharpRulesDefinition class.
